재고감소 기능추가 완료, 예외 롤백 진행중

+++ src/main/java/com/hhplus/commerce/_3weeks/api/request/OrderProductsRequest.java
@@ -5,9 +5,9 @@ import lombok.Getter;
 @Getter
 public class OrderProductsRequest {
     private Long product_id;
-    private int buy_count;
+    private Long buy_count;
 
-    public OrderProductsRequest(Long product_id, int buy_count) {
+    public OrderProductsRequest(Long product_id, Long buy_count) {
         if (buy_count < 0 || buy_count == 0) {
             throw new IllegalStateException("상품은 1개 이상 구매 가능합니다.");
         }
